Rename bokeh_demo -> viz_tool
