+++ .github/workflows/python-app.yaml
@@ -35,8 +35,8 @@ jobs:
       run: cd matfact && poetry install
     - name: Poetry in HMM
       run: cd hmm_synthetic && poetry install
-    - name: Poetry in bokeh_app
-      run: cd bokeh-demo && poetry install
+    - name: Poetry in viz-tool
+      run: cd viz-tool && poetry install
     - name: Poetry in processing
       run: cd processing && poetry install
     - name: Run pre-commit
