+++ viz-tool/examples/landing_page/static/images/SC_logo.png
