+++ viz-tool/examples/pilot/static/images/SC_logo.png
